jas_stream: use `long` instead of `ssize_t`
In the `jas_stream_t` API, offsets are `long`, so why bother using
`ssize_t` here.

This allows removing the `ssize_t` macro from `jas_config.h.in`, which
could break applications using JasPer.

+++ src/libjasper/base/jas_stream.c
@@ -1340,7 +1340,7 @@ static int mem_write(jas_stream_obj_t *obj, const char *buf, unsigned cnt)
 static long mem_seek(jas_stream_obj_t *obj, long offset, int origin)
 {
 	jas_stream_memobj_t *m = (jas_stream_memobj_t *)obj;
-	ssize_t newpos;
+	long newpos;
 
 	JAS_DBGLOG(100, ("mem_seek(%p, %ld, %d)\n", obj, offset, origin));
 	switch (origin) {
